Improve docs Chapter03_02 [ci skip]

+++ code_snippets/chapter03/chapter03_02-002a_explore_prime_numbers.cpp
@@ -1,3 +1,10 @@
+///////////////////////////////////////////////////////////////////////////////
+//  Copyright Christopher Kormanyos 2017 - 2023.
+//  Distributed under the Boost Software License,
+//  Version 1.0. (See accompanying file LICENSE_1_0.txt
+//  or copy at http://www.boost.org/LICENSE_1_0.txt)
+//
+
 #include <algorithm>
 #include <cmath>
 #include <cstdint>
@@ -14,7 +21,7 @@
 namespace
 {
   template<typename unsigned_integral_type = std::uint32_t,
-           const unsigned_integral_type maximum_value = UINT32_C(10006730)>
+           const unsigned_integral_type maximum_value = static_cast<std::uint32_t>(UINT32_C(10006730))>
   void compute_primes_via_sieve(std::vector<unsigned_integral_type>& primes)
   {
     // Use a sieve algorithm to generate
